+++ src/sbatch/mod.rs
@@ -8,6 +8,7 @@ pub mod error;
 pub mod handle;
 pub mod manager;
 pub mod parse;
+pub(crate) mod qgroup_cache;
 pub mod store;
 
 #[cfg(test)]